Rename `nativeFunctions`
