+++ src/reanimated2/functions/index.ts
@@ -1,0 +1,4 @@
+export { dispatchCommand } from './dispatchCommand';
+export { measure } from './measure';
+export { scrollTo } from './scrollTo';
+export { setGestureState } from './setGestureState';
